requirements.txt
Django==4.2.7
mysqlclient==2.2.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
Pillow==10.1.0
whitenoise==6.6.0